move controller logic from experiments to experiments.list
